+++ package.json
@@ -25,8 +25,9 @@
     "typescript": "^3.8.3"
   },
   "scripts": {
-    "deps": "lerna exec yarn",
     "build": "lerna run build",
+    "clean": "lerna clean",
+    "deps": "lerna exec yarn",
     "lint": "lerna run lint"
   },
   "husky": {
@@ -34,10 +35,5 @@
       "pre-commit": "git add . && lerna run lint",
       "pre-push": "lerna run build"
     }
-  },
-  "dependencies": {
-    "@darwinia/api": "file:packages/api",
-    "@darwinia/dj": "file:packages/dj",
-    "@darwinia/util": "file:packages/util"
   }
 }
