+++ packages/api/package.json
@@ -1,7 +1,7 @@
 {
   "private": false,
   "name": "@darwinia/api",
-  "version": "0.1.6",
+  "version": "0.1.5",
   "description": "High-level darwinia.js api",
   "homepage": "https://github.com/darwinia-network/darwinia.js",
   "repository": {
@@ -15,6 +15,7 @@
     "lib/**/*"
   ],
   "dependencies": {
+    "@darwinia/util": "^0.1.5",
     "@polkadot/api": "^1.10.1",
     "@polkadot/keyring": "^2.7.1",
     "@polkadot/util-crypto": "^2.7.1",
