+++ packages/dj/package.json
@@ -1,7 +1,7 @@
 {
   "private": false,
   "name": "@darwinia/dj",
-  "version": "0.1.6",
+  "version": "0.1.5",
   "description": "High-level darwinia.js api",
   "homepage": "https://github.com/darwinia-network/darwinia.js",
   "repository": {
@@ -18,8 +18,8 @@
     "lib/**/*"
   ],
   "dependencies": {
-    "@darwinia/api": "file:../api",
-    "@darwinia/util": "file:../util",
+    "@darwinia/api": "0.1.5",
+    "@darwinia/util": "0.1.5",
     "chalk": "^4.0.0",
     "knex": "^0.20.14",
     "sqlite3": "^4.1.1",
