+++ packages/util/package.json
@@ -1,6 +1,6 @@
 {
  "name": "@darwinia/util",
- "version": "0.1.6",
+ "version": "0.1.5",
  "description": "Util for developing darwinia javascript library.",
  "homepage": "https://github.com/darwinia-network/darwinia.js",
  "repository": {
@@ -21,7 +21,7 @@
   "progress": "^2.0.3"
  },
  "scripts": {
-  "build": "tsc --strict",
+  "build": "tsc --strict && cp src/static/grammer.yml lib/src/static/grammer.yml",
   "lint": "tsc --noEmit --strict && tslint --project ./tsconfig.json"
  }
 }
